Integrate the obstacle avoidance function into the whole motion script.
