+++ jsk_uav_forest_simulation/launch/forest_simulation.launch
@@ -9,10 +9,10 @@
   <arg name="task_kind" default="2" />
   <arg name="rqt_gui" default="True" />
   <arg name="circle_motion_times" default="1" />
-  <arg name="plannar" default="True" />
+  <arg name="do_avoidance" default="True" />
   <arg name="turn_before_return" default="False" />
   <arg name="collision" default="false" />
-  <arg name="use_mapping" default="true" />
+  <arg name="do_mapping" default="true" />
 
   <arg name="start_x" value="-8"/>
   <arg name="start_y" value="0" unless="$(arg collision)"/>
@@ -44,12 +44,12 @@
     <arg name="task_kind" value="$(arg task_kind)" />
     <arg name="circle_motion_times" value="$(arg circle_motion_times)" />
     <arg name="rqt_gui" value="$(arg rqt_gui)" />
-    <arg name="motion_planning" value="$(arg plannar)" />
+    <arg name="do_avoidance" value="$(arg do_avoidance)" />
     <arg name="turn_before_return" value="$(arg turn_before_return)" />
   </include>
 
   <!-- 2D Mapping -->
-  <include file="$(find jsk_uav_forest_common)/launch/2d_mapping.launch" if="$(arg use_mapping)"/>
+  <include file="$(find jsk_uav_forest_common)/launch/2d_mapping.launch" if="$(arg do_mapping)"/>
 
   <node pkg="rviz" type="rviz" name="rviz" args="-d $(find jsk_uav_forest_simulation)/launch/simulation.rviz" unless="$(arg headless)" respawn="true"/>
   <node pkg="jsk_uav_forest_simulation" type="uav_teleop_keyboard" name="uav_teleop_keyboard" launch-prefix="xterm -e" if = "$(arg manual)" />
